+++ e2e-cypress/files/tests/integration/integration.spec.cy.ts
@@ -1,4 +1,4 @@
 /* tslint:disable:no-unused-expression */
 describe('integration e2e tests', () => {
-  // see installation.spec.ts for examples
+  // see acceptance.spec.cy.ts for examples
 });